Update and rename lerpScoreView.cs to LerpScoreView.cs
